Format Params as they're added to the resource
For all values other than email and phone, the required format for the keys is very strange. I thought that this could be one way to improve this, or at least put more of this weirdness under the hood.

Happy to update or get shot down if this isn't going to work, just an idea.